requirements.txt
Django==3.2.12
django-simple-captcha>=0.5.12
django-rq==2.5.1
ldap3==2.9.1
redis==4.4.0
social-auth-core>=4.4.2
social-auth-app-django==5.0.0
python-jose==3.3.0
python3-openid==3.2.0
paramiko==3.1.0
mysqldb
beautifulsoup4
